feat(auth): add authorization
Authorization works in strange ways. Since through `RestFramework` we
set `Access-Control-Allow-Origin: *` we are not allowed to pass
`credentials` on request, which means we cannot pass `Cookies` with
`sessionid`.

The solution approach is to add a new *AuthSystem* on Backend so user
can gets itself from there.

Changes to be committed:
	new file:   .env.example
	modified:   .eslintrc.js
	modified:   config/webpack/webpack.common.js
	modified:   package-lock.json
	modified:   package.json
	modified:   src/components/App.tsx
	new file:   src/const/globalConst.ts
	new file:   src/contexts.tsx
	modified:   tsconfig.json